Catch Postgres error from pg_notify (#74)
The pg_notify payload size is limited to 8000 bytes, the payload should really only contains row IDs but Blockscout puts block & tx data in it, which seems to frequently exceed the max payload size. When the max payload size is exceeded pg_notify throws an exception which makes the indexer process blow up.

Until this misuse of pg_notify is addressed upstream we'll just suppress the exception thrown by it to make the indexer more resilient.